requirements.txt
alembic==1.7.7
black==22.3.0
Bootstrap-Flask==2.0.2
click==8.1.3
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.36
tomli==2.0.1
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
